changes to the POST endpoints and adding the json examples for the objects
