+++ jsons/user.json
@@ -1,0 +1,7 @@
+{
+    "_id": "shalala",
+    "wallet": "wallet_address",
+    "nickname": "user_nick_name",
+    "created_at": "2018-01-01 00:00:00",
+    "updated_at": "2018-01-01 00:00:00"
+}
