+++ pages/api/dream/[dream]/index.js
@@ -15,28 +15,7 @@ export default async function handler(req, res) {
                 const result = await dreams.find({id:dream}).toArray();
                 res.status(200).json({ dream: result });
             }
-            break;
-        case "POST":
-            // Create a new dream project
-            {
-                const date = new Date();
-                const {dream} = query;
-                const newDream = { 
-                    user_id: body.user_id, //id del objeto del usuario
-                    title: body.nickname, // nombre del proyecto
-                    description: body.description, //descripcion del proyecto
-                    main_goal: body.main_goal, //meta de soles del proyecto
-                    benefits: body.benefits, // beneficios del proyecto
-                    deadline: body.deadline, // fecha limite del proyecto
-                    nft_prompt: body.nft_prompt, // prompt del nft
-                    created_at: date, // fecha de creacion del proyecto
-                    updated_at: date, // fecha de actualizacion del proyecto
-                }
-    
-                const result = await dreams.insertOne(newDream);
-                res.status(201).json({message:"Dream created", profile: result.ops});
-            }            
-            break;
+            break;        
         case "PUT":
             // Update the users dream project information
             {
